requirements.txt
Flask>=2
werkzeug>=2
Jinja2==3.1.2
flask-session>=0.3.2,<0.5
requests>=2,<3
# pymssql==2.2.7
Flask-SQLAlchemy==3.0.3
Flask-WTF
python-dotenv==1.0.0
# pymysql==1.0.3
bleach
cryptography
sqlacodegen-v2
pyodbc
mysqlclient
flask-bcrypt 
flask-login 
flask-wtf